Isolated navigator file for future escale
