+++ src/infrastructure/navigation/index.js
@@ -1,0 +1,6 @@
+import React from "react";
+import { AppNavigator } from "./app.navigator";
+
+export const Navigation = () => {
+  return <AppNavigator />;
+};
